使用autocast修复fp16结果, 以及qps对比
