+++ runtime/triton_trtllm/cosyvoice3_trt_fp16_fixed/qps_rslt.md
@@ -1,0 +1,7 @@
+GTX4090 with stream1:
+
+trt_fp16_org:       5.4376 ms, 183.90 qps, with nan.
+
+trt_fp16_cus_ln:    6.0130 ms, 166.22 qps, no nan.
+
+trt_fp16_autocast:  7.1172 ms, 140.45 qps, no nan.
