+++ runtime/triton_trtllm/cosyvoice3_trt_fp16_fixed/v1_plugin/plugin_ln_3d_eps6/CMakeLists.txt
